Update to bflite.md
Fixed typo in prefLabel. Added dateBirth and dateDeath as refinements of date (equal to schema.org). Added e-mail as property of agent. Thought about adding address, but decided that might need to be a refinement of Place and needs more discussion (Place is currently used as a placeholder for Agent profiles for BIBFLOW). Made mark-down spacing consistent. Removed indefinite articles and periods from descriptions for consistency (a request made by Nancy, Jackie, and Xaioli).